Replace `foldr (\a b -> a ++ " " ++ b) ""` with `unwords`
Strictly speaking, the behaviour of the foldr and unwords is not exactly
the same. The former produces `"x y "` out of `["x" "y"]` while the latter
produces `"x y"`. But in this particular case it doesn't change the end result.
In other words, this commit doesn't change the user-observable behaviour of the app.